拆分model文件 去除umi ui
